@@ -7,6 +7,8 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace FSi\Bundle\DoctrineExtensionsBundle\Listener\Uploadable\FileHandler;
 
 use FSi\DoctrineExtensions\Uploadable\Exception\RuntimeException;
@@ -16,26 +18,17 @@ use FSi\Bundle\DoctrineExtensionsBundle\Exception\Uploadable\InvalidFileExceptio
 
 class SymfonyUploadedFileHandler extends AbstractHandler
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function supports($file)
+    public function supports($file): bool
     {
         return $file instanceof UploadedFile;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getName($file)
+    public function getName($file): string
     {
         return $file->getClientOriginalName();
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getContent($file)
+    public function getContent($file): string
     {
         if (!$this->supports($file)) {
             throw $this->generateNotSupportedException($file);
